Rewrite as library
Having gocyclo.Analyze func available makes it possible
for developers to integrate complexity validation
into their tests.
