+++ cmd/gocyclo/main_test.go
@@ -1,0 +1,10 @@
+package main_test
+
+import (
+	"os/exec"
+	"testing"
+)
+
+func TestHelp(t *testing.T) {
+	exec.Command("gocycle", "-h").Run()
+}
